+++ example/lucifer_example.dart
@@ -9,6 +9,7 @@ void main() async {
   });
 
   await app.listen(port);
+
   print('Server running at http://${app.host}:${app.port}');
   app.checkRoutes();
 }