+++ example/simple/lib/middleware/custom.dart
@@ -1,0 +1,7 @@
+import 'package:lucifer/lucifer.dart';
+
+Callback custom() {
+  return (Req req, Res res) async {
+    // define your middleware process here
+  };
+}